Got rooms working
I...kind of hate using the editor; the issue is that, unfortunately, the
html file is nigh unreadable, the code is impossible to organize, and
there is no syntax highlighting. I guess it's an exercise in old-school
coding but, truthfully, I could do just as well without!

Maybe I should check out tweego or summin'.